+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 2.14.0 (2023-09-28)
 
 * (theimo1221) Update packages
 * (theimo1221) Rebase Automatic Points into HeatGroup Settings